refactor: update definition and usages
remove types from services and genesisStore and update usages in whole app
