+++ src/components/ProposalCard.tsx
@@ -3,8 +3,8 @@ import { useMemo } from 'react';
 
 import Tooltip from '@/components/Tooltip';
 import { DAO_UNITS } from '@/config';
-import type { ProposalDetail } from '@/services/proposals';
 import useGenesisStore from '@/stores/genesisStore';
+import type { ProposalDetail } from '@/types/proposal';
 import { getProposalEndTime, uiTokens } from '@/utils';
 
 import { TransactionBadge } from './TransactionBadge';
