+++ src/services/proposals.ts
@@ -1,77 +1,14 @@
 import { BN } from '@polkadot/util';
 
 import { SERVICE_URL } from '@/config';
+import type { IncomingProposal, ProposalDetail } from '@/types/proposal';
+import { proposalStatusNames } from '@/types/proposal';
 import type { Paginated } from '@/types/response';
 import { camelToSnakeCase } from '@/utils';
-
-export enum ProposalStatus {
-  Active = 'Active',
-  Counting = 'Counting',
-  Accepted = 'Accepted',
-  Rejected = 'Rejected',
-  Faulty = 'Faulty',
-}
-
-export type ProposalStatusNames = {
-  [key: string]: ProposalStatus;
-};
-
-export const proposalStatusNames: ProposalStatusNames = {
-  RUNNING: ProposalStatus.Active,
-  PENDING: ProposalStatus.Counting,
-  REJECTED: ProposalStatus.Rejected,
-  IMPLEMENTED: ProposalStatus.Accepted,
-  FAULTED: ProposalStatus.Faulty,
-};
-
-export interface ProposalDetail {
-  proposalId: string;
-  daoId: string;
-  creator: string;
-  birthBlock: number;
-  metadataUrl: string | null;
-  metadataHash: string | null;
-  status: ProposalStatus | null;
-  inFavor: BN;
-  against: BN;
-  voterCount: BN;
-  proposalName: string | null;
-  description: string | null;
-  link: string | null;
-  metadata?: null | {
-    url: string | null;
-    title: string | null;
-    description: string | null;
-  };
-  setupComplete: boolean;
-}
 
 export interface ReportFaultedRequestPayload {
   proposalId: string;
   reason: string;
-}
-
-export interface IncomingProposal {
-  id: string;
-  dao_id: string;
-  creator_id: string;
-  status: string;
-  fault: string | null;
-  votes: {
-    pro: number;
-    contra: number;
-    abstained: number;
-    total: number;
-  };
-  metadata: null | {
-    url: string | null;
-    title: string | null;
-    description: string | null;
-  };
-  metadata_url: null | string;
-  metadata_hash: null | string;
-  birth_block_number: number;
-  setup_complete: boolean;
 }
 
 export interface ListProposalsQueryParams {
